+++ replacer-backend/src/main/java/es/bvalero/replacer/review/dto/ReviewMapper.java
@@ -104,6 +104,7 @@ public class ReviewMapper {
             .cs(reviewed.getCs())
             .start(reviewed.getStart())
             .reviewer(queryParameters.getUser())
+            .fixed(reviewed.isFixed())
             .build();
     }
 }