+++ replacer-backend/src/main/java/es/bvalero/replacer/review/save/ReviewedReplacement.java
@@ -24,4 +24,6 @@ public class ReviewedReplacement {
 
     @NonNull
     String reviewer;
+
+    boolean fixed;
 }